+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+* Fix python command recognition: avoid wrapping for -c module argument
+
 ### 1.0.0 (2017-12-20)
 * Initial release